+++ Sources/StreamChat/WebSocketClient/WebSocketConnectPayload.swift
@@ -29,6 +29,7 @@ struct UserWebSocketPayload: Encodable {
         case isInvisible = "invisible"
         case imageURL = "image"
         case language
+        case privacySettings = "privacy_settings"
     }
 
     let id: String
@@ -36,6 +37,7 @@ struct UserWebSocketPayload: Encodable {
     let imageURL: URL?
     let isInvisible: Bool?
     let language: String?
+    let privacySettings: UserPrivacySettingsPayload?
     let extraData: [String: RawJSON]
 
     init(userInfo: UserInfo) {
@@ -44,6 +46,7 @@ struct UserWebSocketPayload: Encodable {
         imageURL = userInfo.imageURL
         isInvisible = userInfo.isInvisible
         language = userInfo.language?.languageCode
+        privacySettings = userInfo.privacySettings.map { .init(settings: $0) }
         extraData = userInfo.extraData
     }
 
@@ -54,6 +57,44 @@ struct UserWebSocketPayload: Encodable {
         try container.encodeIfPresent(imageURL, forKey: .imageURL)
         try container.encodeIfPresent(isInvisible, forKey: .isInvisible)
         try container.encodeIfPresent(language, forKey: .language)
+        try container.encodeIfPresent(privacySettings, forKey: .privacySettings)
         try extraData.encode(to: encoder)
     }
+}
+
+struct UserPrivacySettingsPayload: Codable {
+    enum CodingKeys: String, CodingKey {
+        case typingIndicators = "typing_indicators"
+        case readReceipts = "read_receipts"
+    }
+
+    let typingIndicators: TypingIndicatorPrivacySettingsPayload?
+    let readReceipts: ReadReceiptsPrivacySettingsPayload?
+
+    init(
+        typingIndicators: TypingIndicatorPrivacySettingsPayload?,
+        readReceipts: ReadReceiptsPrivacySettingsPayload?
+    ) {
+        self.typingIndicators = typingIndicators
+        self.readReceipts = readReceipts
+    }
+
+    init(settings: UserPrivacySettings) {
+        typingIndicators = settings.typingIndicators.map { .init(enabled: $0.enabled) }
+        readReceipts = settings.readReceipts.map { .init(enabled: $0.enabled) }
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encodeIfPresent(typingIndicators, forKey: .typingIndicators)
+        try container.encodeIfPresent(readReceipts, forKey: .readReceipts)
+    }
+}
+
+struct TypingIndicatorPrivacySettingsPayload: Codable {
+    var enabled: Bool
+}
+
+struct ReadReceiptsPrivacySettingsPayload: Codable {
+    var enabled: Bool
 }