+++ Sources/StreamChat/Workers/CurrentUserUpdater.swift
@@ -15,12 +15,14 @@ class CurrentUserUpdater: Worker {
     ///   - currentUserId: The current user identifier.
     ///   - name: Optionally provide a new name to be updated.
     ///   - imageURL: Optionally provide a new image to be updated.
+    ///   - privacySettings: The privacy settings of the user. Example: If the user does not want to expose typing events or read events.
     ///   - userExtraData: Optionally provide new user extra data to be updated.
     ///   - completion: Called when user is successfuly updated, or with error.
     func updateUserData(
         currentUserId: UserId,
         name: String? = nil,
         imageURL: URL? = nil,
+        privacySettings: UserPrivacySettings? = nil,
         userExtraData: [String: RawJSON]? = nil,
         completion: ((Error?) -> Void)? = nil
     ) {
@@ -34,6 +36,7 @@ class CurrentUserUpdater: Worker {
         let payload = UserUpdateRequestBody(
             name: name,
             imageURL: imageURL,
+            privacySettings: privacySettings.map { UserPrivacySettingsPayload(settings: $0) },
             extraData: userExtraData
         )
 
@@ -42,8 +45,7 @@ class CurrentUserUpdater: Worker {
                 switch $0 {
                 case let .success(response):
                     self?.database.write({ (session) in
-                        let userDTO = try session.saveUser(payload: response.user)
-                        session.currentUser?.user = userDTO
+                        try session.saveCurrentUser(payload: response.user)
                     }) { completion?($0) }
                 case let .failure(error):
                     completion?(error)