+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## StreamChat
 ### ✅ Added
+- Add user privacy settings to control user events [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
+- Add `privacySettings` parameter to `connectUser()` and `CurrentUserController.updateUserData()` [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
+- Expose `ChatCurrentUser.privacySettings` [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
 - Adds `reactionGroups` to `ChatMessage` so that reactions can be sorted by date in the message list [#3158](https://github.com/GetStream/stream-chat-swift/pull/3158)
 
 ## StreamChatUI